+++ contracts/interfaces/IW12CrowdsaleFactory.sol
@@ -4,5 +4,5 @@ import "./IW12Crowdsale.sol";
 
 
 interface IW12CrowdsaleFactory {
-    function createCrowdsale(address _wTokenAddress, uint32 _startDate, uint price, address serviceWallet, uint8 serviceFee, address swap, address owner) external returns (IW12Crowdsale);
+    function createCrowdsale(address _wTokenAddress, uint32 _startDate, uint price, address serviceWallet, uint serviceFee, address swap, address owner) external returns (IW12Crowdsale);
 }
